requirements.txt
stripe==14.3.0
requests==2.32.3
python-dotenv==1.0.0